split mapping and sorting
