+++ wls-gui-admintool/src/composables/wahltag/wahltagMapper.ts
@@ -1,6 +1,7 @@
 import type { WahltagDTO } from "@/api/wls-clients/generated-admin-api";
 import type { Wahltag } from "@/types/wahltag/Wahltag.ts";
-import type { WahltagEvent } from "@/types/wahltag/WahltagEvent.ts";
+
+import { type WahltagEvent } from "@/types/wahltag/WahltagEvent.ts";
 
 export function useWahltagMapper() {
   function mapWahltagDtoToWahltagEvent(dto: WahltagDTO): WahltagEvent {
@@ -20,8 +21,28 @@ export function useWahltagMapper() {
     };
   }
 
+  function mapGroupedWahltagDtosToWahltage(
+    groupedDtos: Map<string, WahltagDTO[]>
+  ): Wahltag[] {
+    const result: Wahltag[] = [];
+
+    groupedDtos.forEach((wahltage, wahltagDatum) => {
+      const wahltagEvents = wahltage.map((dto) =>
+        mapWahltagDtoToWahltagEvent(dto)
+      );
+
+      result.push({
+        wahltag: wahltagDatum,
+        events: wahltagEvents,
+      });
+    });
+
+    return result;
+  }
+
   return {
     wahltagModelToWahltagDto: mapWahltagModelToWahltagDto,
     mapWahltagDtoToWahltagEvent,
+    mapGroupedWahltagDtosToWahltage,
   };
 }
